requirements.txt
chromadb==0.5.23
Flask==3.1.0
langchain_community==0.3.19
langchain_openai==0.3.8
langchain_text_splitters==0.3.6
matplotlib==3.10.1
numpy==2.2.3
openai==1.65.5
pandas==2.2.3
python-dotenv==1.0.1
qdrant_client==1.13.3
ragas==0.2.14
Requests==2.32.3
scikit_learn==1.6.1
